requirements.txt
Flask==2.3.2
transformers==4.28.1
opencv-python==4.7.0.72
numpy==1.24.2
python-doctr==0.6.0
jinja2==3.1.2

